test: fix QuackGraph import in temporal tests and add db assertions to persistence test
fix: ensure temporal columns hydrate correctly by casting to DOUBLE
Casting valid_from/valid_to to DOUBLE in the hydration SQL query prevents data loss when the Arrow IPC fallback path (tableFromJSON) encounters BigInts. This resolves issues where soft-deleted edges appeared active and future edges appeared in the past.

+++ packages/quack-graph/src/graph.ts
@@ -90,12 +90,12 @@ export class QuackGraph {
   async hydrate() {
     // Zero-Copy Arrow IPC
     // We load ALL edges (active and historical) to support time-travel.
-    // We cast valid_from/valid_to to BIGINT (INT64) to ensure Arrow compatibility
+    // We cast valid_from/valid_to to DOUBLE to ensure JS/JSON compatibility (avoiding BigInt issues in fallback)
     try {
       const ipcBuffer = await this.db.queryArrow(
         `SELECT source, target, type, 
-                date_diff('us', '1970-01-01'::TIMESTAMPTZ, valid_from) as valid_from, 
-                date_diff('us', '1970-01-01'::TIMESTAMPTZ, valid_to) as valid_to 
+                date_diff('us', '1970-01-01'::TIMESTAMPTZ, valid_from)::DOUBLE as valid_from, 
+                date_diff('us', '1970-01-01'::TIMESTAMPTZ, valid_to)::DOUBLE as valid_to 
          FROM edges`
       );
